#965 Introduce bridged buses, same concept as bridged quadripoles
