+++ dynawo/sources/Models/CPP/ModelNetwork/DYNModelNetwork.cpp
@@ -35,8 +35,11 @@
 #include "PARParametersSet.h"
 #include "PARParametersSetFactory.h"
 #include "DYNModelConstants.h"
-#include "DYNModelBus.h"
+#include "DYNModelBusInjected.h"
+#include "DYNModelBusBridged.h"
+#include "DYNModelBusContainer.h"
 #include "DYNModelSwitchFactory.h"
+#include "DYNModelSubNetwork.hpp"
 #include "DYNModelLine.h"
 #include "DYNModelTwoWindingsTransformer.h"
 #include "DYNModelThreeWindingsTransformer.h"
@@ -171,23 +174,29 @@ ModelNetwork::initializeFromData(const shared_ptr<DataInterface>& data) {
     // ==============================
     for (const auto& bus : voltageLevel->getBuses()) {
       string id = bus->getID();
-      std::shared_ptr<ModelBus> modelBus(new ModelBus(bus, voltageLevel->isNodeBreakerTopology()));
       componentsById[id] = bus;
-      modelBusById[id] = modelBus;
-      // Add to containers
-      modelVoltageLevelInit->addBus(modelBus);
+      std::shared_ptr<ModelBus> modelBus;
+      if (bus->hasDynamicModel()) {
+        std::shared_ptr<ModelBusBridged> modelBusBridged(new ModelBusBridged(bus, voltageLevel->isNodeBreakerTopology()));
+        unmappedBridges_[id] = modelBusBridged;
+        modelBus = modelBusBridged;
+        Trace::debug(Trace::network()) << DYNLog(BusExtDynModel, id) << Trace::endline;
+      } else {
+        modelBus = std::shared_ptr<ModelBusInjected>(new ModelBusInjected(bus, voltageLevel->isNodeBreakerTopology()));
+        // declare reference between subModel and static data
+        data->setReference("v", id, id, "U_value");
+        data->setReference("angle", id, id, "phi_value");
+        Trace::debug(Trace::network()) << DYNLog(AddingBusToNetwork, id) << Trace::endline;
+      }
+
       modelBus->setNetwork(this);
       modelBus->setVoltageLevel(modelVoltageLevel);
-      if (bus->hasDynamicModel()) {
-        Trace::debug(Trace::network()) << DYNLog(BusExtDynModel, id) << Trace::endline;
-        continue;
-      }
-      Trace::debug(Trace::network()) << DYNLog(AddingBusToNetwork, id) << Trace::endline;
+
+      // Add to containers
+      modelBusById[id] = modelBus;
+      modelVoltageLevelInit->addBus(modelBus);
       modelVoltageLevel->addBus(modelBus);
       busContainer_->add(modelBus);
-      // declare reference between subModel and static data
-      data->setReference("v", id, id, "U_value");
-      data->setReference("angle", id, id, "phi_value");
     }
 
     // =============================
@@ -368,7 +377,10 @@ ModelNetwork::initializeFromData(const shared_ptr<DataInterface>& data) {
     initComponents_.push_back(modelLine);
 
     if (line->hasDynamicModel()) {
-      addBridge(std::make_shared<NetworkBridgeQuadripole>(modelLine, "line", this));
+      std::shared_ptr<NetworkBridgeQuadripole> bridge = std::make_shared<NetworkBridgeQuadripole>(modelLine, "line");
+      bridge->setNetwork(this);
+      components_.push_back(bridge);
+      unmappedBridges_[id] = bridge;
       Trace::debug(Trace::network()) << DYNLog(LineExtDynModel, id) << Trace::endline;
       continue;
     }
@@ -408,9 +420,11 @@ ModelNetwork::initializeFromData(const shared_ptr<DataInterface>& data) {
 
     initComponents_.push_back(modelTwoWindingsTransformer);
 
-
     if (twoWTfo->hasDynamicModel()) {
-      addBridge(std::make_shared<NetworkBridgeQuadripole>(modelTwoWindingsTransformer, "transformer", this));
+      std::shared_ptr<NetworkBridgeQuadripole> bridge = std::make_shared<NetworkBridgeQuadripole>(modelTwoWindingsTransformer, "transformer");
+      bridge->setNetwork(this);
+      components_.push_back(bridge);
+      unmappedBridges_[id] = bridge;
       Trace::debug(Trace::network()) << DYNLog(TwoWTfoExtDynModel, id) << Trace::endline;
       continue;
     }
@@ -1082,11 +1096,11 @@ ModelNetwork::evalJt(const double /*t*/, const double cj, const int rowOffset, S
   Timer timer("ModelNetwork::evalJ");
 #endif
 
-  // init bus derivatives
+  // reset bus derivatives
 #if defined(_DEBUG_) || defined(PRINT_TIMERS)
-  Timer* timer2 = new Timer("ModelNetwork::evalJt_initBusDerivatives");
+  Timer* timer2 = new Timer("ModelNetwork::evalJt_resettBusDerivatives");
 #endif
-  busContainer_->initDerivatives();
+  busContainer_->resetDerivatives();
 #if defined(_DEBUG_) || defined(PRINT_TIMERS)
   delete timer2;
 #endif
@@ -1494,20 +1508,11 @@ ModelNetwork::loadInternalVariables(boost::archive::binary_iarchive&) {
 }
 
 void
-ModelNetwork::addBridge(const std::shared_ptr<NetworkBridgeQuadripole> & bridge) {
-  components_.push_back(bridge);
-  unmappedBridges_[bridge->id()] = bridge;
-}
-
-void
 ModelNetwork::mapToNetworkBridge(const boost::shared_ptr<SubModel> & submodel) {
-  std::string bridgeId = NetworkBridgeQuadripole::BRIDGE_PREFIX + submodel->staticId();
-
-  if (unmappedBridges_.find(bridgeId) != unmappedBridges_.end()) {
-    unmappedBridges_[bridgeId]->setDynPart(submodel);
-    unmappedBridges_.erase(bridgeId);
-  }
+  if (unmappedBridges_.find(submodel->staticId()) == unmappedBridges_.end())
+    return;
+  unmappedBridges_[submodel->staticId()]->setDynPart(submodel);
+  unmappedBridges_.erase(submodel->staticId());
 }
-
 
 }  // namespace DYN
