+++ dynawo/sources/Models/CPP/ModelNetwork/DYNModelNetwork.h
@@ -34,7 +34,7 @@ class ModelBusContainer;
 class ModelSwitch;
 class ModelVoltageLevel;
 class NetworkComponent;
-class NetworkBridgeQuadripole;
+class NetworkBridge;
 class DataInterface;
 
 static const double maximumValueCurrentLimit = 5000;   ///< Maximum acceptable value for current limits
@@ -403,12 +403,6 @@ class ModelNetwork : public ModelCPP, private boost::noncopyable {
   */
   void printInternalParameters(std::ofstream& fstream) const override;
 
-  /**
-  * @brief add a newly instanciated NetworkBridge to adequate containers
-  * @param bridge said newly instanciated NetworkBridge
-  */
-  void addBridge(const std::shared_ptr<NetworkBridgeQuadripole> & bridge);
-
  private:
   double* calculatedVarBuffer_;  ///< calculated variable buffer
 
@@ -421,7 +415,7 @@ class ModelNetwork : public ModelCPP, private boost::noncopyable {
   std::vector<std::shared_ptr<ModelVoltageLevel> > vLevelInitComponents_;  ///< all voltage level components  (used for init model)
   std::vector<std::shared_ptr<NetworkComponent> > components_;  ///< all network components without dynamic Model
   std::vector<std::shared_ptr<NetworkComponent> > initComponents_;  ///< all network components even components with dynamic model
-  std::unordered_map<std::string, std::shared_ptr<NetworkBridgeQuadripole> > unmappedBridges_;  ///< network bridges yet unassociated with their dynamic model
+  std::unordered_map<std::string, std::shared_ptr<NetworkBridge> > unmappedBridges_;  ///< network bridges yet unassociated with their dynamic model
   std::vector<int> componentIndexByCalculatedVar_;  ///< index of component for each calculated variable
 };
