+++ dynawo/sources/Models/CPP/ModelNetwork/test/TestLine.cpp
@@ -27,7 +27,7 @@
 #include "DYNBusInterfaceIIDM.h"
 #include "DYNModelLine.h"
 #include "DYNModelVoltageLevel.h"
-#include "DYNModelBus.h"
+#include "DYNModelBusInjected.h"
 #include "DYNModelNetwork.h"
 #include "TLTimelineFactory.h"
 #include "DYNSparseMatrix.h"
@@ -153,7 +153,7 @@ createModelLine(bool open, bool initModel, bool closed1 = true, bool closed2 = t
   std::shared_ptr<ModelVoltageLevel> vl = std::make_shared<ModelVoltageLevel>(vlItfIIDM);
   int offset = 0;
 
-  std::shared_ptr<ModelBus> bus1 = std::make_shared<ModelBus>(bus1ItfIIDM, false);
+  std::shared_ptr<ModelBusInjected> bus1 = std::make_shared<ModelBusInjected>(bus1ItfIIDM, false);
   bus1->setNetwork(network);
   bus1->setVoltageLevel(vl);
   dl->setModelBus1(bus1);
@@ -168,13 +168,13 @@ createModelLine(bool open, bool initModel, bool closed1 = true, bool closed2 = t
     zConnected1[i] = closed1;
   bus1->setReferenceZ(&z1[0], zConnected1, 0);
   bus1->setReferenceY(y1, yp1, f1, 0, 0);
-  y1[ModelBus::urNum_] = 3.5;
-  y1[ModelBus::uiNum_] = 2;
+  y1[ModelBusInjected::urNum_] = 3.5;
+  y1[ModelBusInjected::uiNum_] = 2;
   if (!initModel)
     z1[ModelBus::switchOffNum_] = -1;
   bus1->init(offset);
 
-  std::shared_ptr<ModelBus> bus2 = std::make_shared<ModelBus>(bus2ItfIIDM, false);
+  std::shared_ptr<ModelBusInjected> bus2 = std::make_shared<ModelBusInjected>(bus2ItfIIDM, false);
   bus2->setNetwork(network);
   bus2->setVoltageLevel(vl);
   dl->setModelBus2(bus2);
@@ -189,8 +189,8 @@ createModelLine(bool open, bool initModel, bool closed1 = true, bool closed2 = t
     zConnected2[i] = closed2;
   bus2->setReferenceZ(&z2[0], zConnected2, 0);
   bus2->setReferenceY(y2, yp2, f2, 0, 0);
-  y2[ModelBus::urNum_] = 4.;
-  y2[ModelBus::uiNum_] = 1.5;
+  y2[ModelBusInjected::urNum_] = 4.;
+  y2[ModelBusInjected::uiNum_] = 1.5;
   if (!initModel)
     z2[ModelBus::switchOffNum_] = -1;
   bus2->init(offset);
