+++ dynawo/sources/Models/CPP/ModelNetwork/test/TestTwoWindingsTransformer.cpp
@@ -31,7 +31,7 @@
 #include "DYNBusInterfaceIIDM.h"
 #include "DYNModelTwoWindingsTransformer.h"
 #include "DYNModelVoltageLevel.h"
-#include "DYNModelBus.h"
+#include "DYNModelBusInjected.h"
 #include "DYNModelNetwork.h"
 #include "TLTimelineFactory.h"
 #include "DYNSparseMatrix.h"
@@ -240,7 +240,7 @@ createModelTwoWindingsTransformer(bool open, bool initModel, bool ratioTapChange
   std::shared_ptr<ModelVoltageLevel> vl = std::make_shared<ModelVoltageLevel>(vlItfIIDM);
   int offset = 0;
   if (closed1) {
-    std::unique_ptr<ModelBus> bus1 = DYN::make_unique<ModelBus>(bus1ItfIIDM, false);
+    std::unique_ptr<ModelBusInjected> bus1 = DYN::make_unique<ModelBusInjected>(bus1ItfIIDM, false);
     bus1->setNetwork(network);
     bus1->setVoltageLevel(vl);
     bus1->initSize();
@@ -254,15 +254,15 @@ createModelTwoWindingsTransformer(bool open, bool initModel, bool ratioTapChange
       zConnected1[i] = true;
     bus1->setReferenceZ(&z1[0], zConnected1, 0);
     bus1->setReferenceY(y1, yp1, f1, 0, 0);
-    y1[ModelBus::urNum_] = 3.5;
-    y1[ModelBus::uiNum_] = 2;
+    y1[ModelBusInjected::urNum_] = 3.5;
+    y1[ModelBusInjected::uiNum_] = 2;
     if (!initModel)
       z1[ModelBus::switchOffNum_] = -1;
     bus1->init(offset);
     t2w->setModelBus1(std::move(bus1));
   }
   if (closed2) {
-    std::unique_ptr<ModelBus> bus2 = DYN::make_unique<ModelBus>(bus2ItfIIDM, false);
+    std::unique_ptr<ModelBusInjected> bus2 = DYN::make_unique<ModelBusInjected>(bus2ItfIIDM, false);
     bus2->setNetwork(network);
     bus2->setVoltageLevel(vl);
     bus2->initSize();
@@ -276,8 +276,8 @@ createModelTwoWindingsTransformer(bool open, bool initModel, bool ratioTapChange
       zConnected2[i] = true;
     bus2->setReferenceZ(&z2[0], zConnected2, 0);
     bus2->setReferenceY(y2, yp2, f2, 0, 0);
-    y2[ModelBus::urNum_] = 4.;
-    y2[ModelBus::uiNum_] = 1.5;
+    y2[ModelBusInjected::urNum_] = 4.;
+    y2[ModelBusInjected::uiNum_] = 1.5;
     if (!initModel)
       z2[ModelBus::switchOffNum_] = -1;
     bus2->init(offset);
