+++ dynawo/sources/Models/Modelica/Dynawo/NonElectrical/Logs/ErrorKeys.mo
@@ -226,64 +226,65 @@ encapsulated package ErrorKeys
   final constant Integer UndefinedComponentState = 208;
   final constant Integer UndefinedNominalV = 209;
   final constant Integer UndefinedStep = 210;
-  final constant Integer UnitModelIDSameAsModelName = 211;
-  final constant Integer UnitModelIDSameAsUnitModelName = 212;
-  final constant Integer UnknownAutomatonOutput = 213;
-  final constant Integer UnknownBus = 214;
-  final constant Integer UnknownCalculatedBus = 215;
-  final constant Integer UnknownChannelId = 216;
-  final constant Integer UnknownComponent = 217;
-  final constant Integer UnknownConstraintsExport = 218;
-  final constant Integer UnknownConstraintsStreamFormat = 219;
-  final constant Integer UnknownCurveFile = 220;
-  final constant Integer UnknownCurvesExport = 221;
-  final constant Integer UnknownCurvesStreamFormat = 222;
-  final constant Integer UnknownDydFile = 223;
-  final constant Integer UnknownFinalStateExport = 224;
-  final constant Integer UnknownFinalStateFile = 225;
-  final constant Integer UnknownFinalStateValuesExport = 226;
-  final constant Integer UnknownFinalStateValuesFile = 227;
-  final constant Integer UnknownIidmFile = 228;
-  final constant Integer UnknownInitialStateFile = 229;
-  final constant Integer UnknownModelFile = 230;
-  final constant Integer UnknownModelsDir = 231;
-  final constant Integer UnknownParFile = 232;
-  final constant Integer UnknownParSet = 233;
-  final constant Integer UnknownStateVariable = 234;
-  final constant Integer UnknownStaticComponent = 235;
-  final constant Integer UnknownStaticParameter = 236;
-  final constant Integer UnknownTimelineExport = 237;
-  final constant Integer UnknownTimelineStreamFormat = 238;
-  final constant Integer UnknownVertex = 239;
-  final constant Integer UnknownVoltageLevel = 240;
-  final constant Integer UnmappedNetworkBridge = 241;
-  final constant Integer UnstableRoots = 242;
-  final constant Integer UnsupportedComponentState = 243;
-  final constant Integer VariableAliasIncoherentType = 244;
-  final constant Integer VariableAliasRefIncoherent = 245;
-  final constant Integer VariableAliasRefNotNative = 246;
-  final constant Integer VariableAliasRefNotSet = 247;
-  final constant Integer VariableCardinalityNotSet = 248;
-  final constant Integer VariableMultipleHasNoIndex = 249;
-  final constant Integer VariableNativeIndexAlreadySet = 250;
-  final constant Integer VariableNativeIndexNotSet = 251;
-  final constant Integer VoltageLevelGraphUndefined = 252;
-  final constant Integer VoltageLevelTopoError = 253;
-  final constant Integer WrongCheckSum = 254;
-  final constant Integer WrongConnect = 255;
-  final constant Integer WrongConnectTwoUnknownNodes = 256;
-  final constant Integer WrongDataNum = 257;
-  final constant Integer WrongDynamicCast = 258;
-  final constant Integer WrongIIDMDataForHVDC = 259;
-  final constant Integer WrongLinearSolverChoice = 260;
-  final constant Integer WrongReferenceId = 261;
-  final constant Integer XercesHandler = 262;
-  final constant Integer XmlFileParsingError = 263;
-  final constant Integer XmlParsingError = 264;
-  final constant Integer XmlUtilsLoadSchema = 265;
-  final constant Integer XmlUtilsXercesInit = 266;
-  final constant Integer ZMQInterfaceBadEnpoint = 267;
-  final constant Integer ZValueIsNaN = 268;
+  final constant Integer UnhandledBridgedBusCall = 211;
+  final constant Integer UnitModelIDSameAsModelName = 212;
+  final constant Integer UnitModelIDSameAsUnitModelName = 213;
+  final constant Integer UnknownAutomatonOutput = 214;
+  final constant Integer UnknownBus = 215;
+  final constant Integer UnknownCalculatedBus = 216;
+  final constant Integer UnknownChannelId = 217;
+  final constant Integer UnknownComponent = 218;
+  final constant Integer UnknownConstraintsExport = 219;
+  final constant Integer UnknownConstraintsStreamFormat = 220;
+  final constant Integer UnknownCurveFile = 221;
+  final constant Integer UnknownCurvesExport = 222;
+  final constant Integer UnknownCurvesStreamFormat = 223;
+  final constant Integer UnknownDydFile = 224;
+  final constant Integer UnknownFinalStateExport = 225;
+  final constant Integer UnknownFinalStateFile = 226;
+  final constant Integer UnknownFinalStateValuesExport = 227;
+  final constant Integer UnknownFinalStateValuesFile = 228;
+  final constant Integer UnknownIidmFile = 229;
+  final constant Integer UnknownInitialStateFile = 230;
+  final constant Integer UnknownModelFile = 231;
+  final constant Integer UnknownModelsDir = 232;
+  final constant Integer UnknownParFile = 233;
+  final constant Integer UnknownParSet = 234;
+  final constant Integer UnknownStateVariable = 235;
+  final constant Integer UnknownStaticComponent = 236;
+  final constant Integer UnknownStaticParameter = 237;
+  final constant Integer UnknownTimelineExport = 238;
+  final constant Integer UnknownTimelineStreamFormat = 239;
+  final constant Integer UnknownVertex = 240;
+  final constant Integer UnknownVoltageLevel = 241;
+  final constant Integer UnmappedNetworkBridge = 242;
+  final constant Integer UnstableRoots = 243;
+  final constant Integer UnsupportedComponentState = 244;
+  final constant Integer VariableAliasIncoherentType = 245;
+  final constant Integer VariableAliasRefIncoherent = 246;
+  final constant Integer VariableAliasRefNotNative = 247;
+  final constant Integer VariableAliasRefNotSet = 248;
+  final constant Integer VariableCardinalityNotSet = 249;
+  final constant Integer VariableMultipleHasNoIndex = 250;
+  final constant Integer VariableNativeIndexAlreadySet = 251;
+  final constant Integer VariableNativeIndexNotSet = 252;
+  final constant Integer VoltageLevelGraphUndefined = 253;
+  final constant Integer VoltageLevelTopoError = 254;
+  final constant Integer WrongCheckSum = 255;
+  final constant Integer WrongConnect = 256;
+  final constant Integer WrongConnectTwoUnknownNodes = 257;
+  final constant Integer WrongDataNum = 258;
+  final constant Integer WrongDynamicCast = 259;
+  final constant Integer WrongIIDMDataForHVDC = 260;
+  final constant Integer WrongLinearSolverChoice = 261;
+  final constant Integer WrongReferenceId = 262;
+  final constant Integer XercesHandler = 263;
+  final constant Integer XmlFileParsingError = 264;
+  final constant Integer XmlParsingError = 265;
+  final constant Integer XmlUtilsLoadSchema = 266;
+  final constant Integer XmlUtilsXercesInit = 267;
+  final constant Integer ZMQInterfaceBadEnpoint = 268;
+  final constant Integer ZValueIsNaN = 269;
 
   annotation(preferredView = "text");
 end ErrorKeys;
